Lang: Move TranslationContainer and TextContainer to \pocketmine\lang namespace why the hell were they ever put in \pocketmine\event in the first place??
This change was suggested many months ago but I forgot all about it.
