+++ src/pocketmine/command/Command.php
@@ -26,9 +26,9 @@ declare(strict_types=1);
  */
 namespace pocketmine\command;
 
-use pocketmine\event\TextContainer;
 use pocketmine\event\TimingsHandler;
-use pocketmine\event\TranslationContainer;
+use pocketmine\lang\TextContainer;
+use pocketmine\lang\TranslationContainer;
 use pocketmine\Server;
 use pocketmine\utils\TextFormat;
