+++ src/pocketmine/command/CommandSender.php
@@ -23,7 +23,7 @@ declare(strict_types=1);
 
 namespace pocketmine\command;
 
-use pocketmine\event\TextContainer;
+use pocketmine\lang\TextContainer;
 use pocketmine\permission\Permissible;
 use pocketmine\Server;
