+++ src/pocketmine/command/defaults/EnchantCommand.php
@@ -25,9 +25,9 @@ namespace pocketmine\command\defaults;
 
 use pocketmine\command\CommandSender;
 use pocketmine\command\utils\InvalidCommandSyntaxException;
-use pocketmine\event\TranslationContainer;
 use pocketmine\item\enchantment\Enchantment;
 use pocketmine\item\enchantment\EnchantmentInstance;
+use pocketmine\lang\TranslationContainer;
 use pocketmine\utils\TextFormat;
 
 class EnchantCommand extends VanillaCommand{
