+++ src/pocketmine/command/defaults/PluginsCommand.php
@@ -24,7 +24,7 @@ declare(strict_types=1);
 namespace pocketmine\command\defaults;
 
 use pocketmine\command\CommandSender;
-use pocketmine\event\TranslationContainer;
+use pocketmine\lang\TranslationContainer;
 use pocketmine\utils\TextFormat;
 
 class PluginsCommand extends VanillaCommand{
