+++ src/pocketmine/command/defaults/SaveOffCommand.php
@@ -25,7 +25,7 @@ namespace pocketmine\command\defaults;
 
 use pocketmine\command\Command;
 use pocketmine\command\CommandSender;
-use pocketmine\event\TranslationContainer;
+use pocketmine\lang\TranslationContainer;
 
 class SaveOffCommand extends VanillaCommand{
