+++ src/pocketmine/command/defaults/SeedCommand.php
@@ -24,7 +24,7 @@ declare(strict_types=1);
 namespace pocketmine\command\defaults;
 
 use pocketmine\command\CommandSender;
-use pocketmine\event\TranslationContainer;
+use pocketmine\lang\TranslationContainer;
 use pocketmine\Player;
 
 class SeedCommand extends VanillaCommand{
