+++ src/pocketmine/event/player/PlayerDeathEvent.php
@@ -24,8 +24,8 @@ declare(strict_types=1);
 namespace pocketmine\event\player;
 
 use pocketmine\event\entity\EntityDeathEvent;
-use pocketmine\event\TextContainer;
 use pocketmine\item\Item;
+use pocketmine\lang\TextContainer;
 use pocketmine\Player;
 
 class PlayerDeathEvent extends EntityDeathEvent{
